backend/requirements.txt
fastapi==0.128.7
uvicorn==0.40.0
SQLAlchemy==2.0.46
psycopg2-binary==2.9.11
passlib==1.7.4
bcrypt==3.2.2
python-jose==3.5.0
pydantic==2.12.5
email-validator==2.3.0
better-profanity==0.7.0
pytest==8.0.0
httpx==0.26.0
python-dotenv==1.0.1
